+++ packages/runtime/src/console/console.ts
@@ -1,0 +1,6 @@
+export interface Console {
+  clear(): void;
+  add(data: string): void;
+  get(): string;
+  getTrimmed(): string;
+}
