+++ packages/runtime/src/context.ts
@@ -1,4 +1,4 @@
-import {Console} from "./console";
+import {Console} from "./console/console";
 import {DatabaseClient} from "./db/db";
 import * as RFC from "./rfc";
